fix(metrics): the bite-list message told you to delete the only guard on `$`
CRITICAL, caught by review. 5ee38f7 rewrote that failure message to be accurate
and got the taxonomy backwards in the one direction that kills the gate.

It called the two TRAILING glosses ("-40% build time", "-97.5% (40s->1s,
Venturus)") "DOMINATED as killers, kept as documentation". They are not
dominated. Recomputed the kill matrix: they are jointly the ONLY two cases that
kill a dropped `$` anchor. Delete both — which the message invites — and the
schema can drop `$`, accept a trailing gloss, and that metric's sweep becomes a
permanent silent no-op. Bite-proved: trim those two and the drop-`$` mutant
SURVIVES with the suite green.

The only strictly dominated case is the LEADING gloss "build time -40%"
(" -40%" kills everything it kills). I had the two glosses backwards. The
message now states the matrix: which cases are sole killers, that the two
trailing glosses are joint-and-only guards of `$` so at least one must stay,
and which single case is dominated.

This is the third time in this branch the same error class landed in the failure
message rather than the code, and the worst of the three: a gate whose message
instructs the engineer to reopen the exact fail-open the gate exists to close.

Also fixed, same review:

- PROJECT_STAT_METRICS is an inclusion list keyed on DISPLAY COPY, and an
  unmapped label is skipped, not failed. This branch itself renamed "TTI GAIN"
  to "TTI", proving the labels move. Rename "JS BUNDLE" tomorrow and that tile
  silently leaves the magnitude gate — the one thing the map exists to hold. New
  assertion reds on an orphaned key; bite-proved by renaming a live label.

- The predicate was DUPLICATED across the two test files, each with its own
  table. Adding an exclusion to one copy diverged the two sweeps with both tables
  green — the same fail-open class again. Extracted to
  __tests__/helpers/content-surfaces.ts: one predicate, one table. This also
  makes the ADR's "shares isPublishedSurface" claim true; it previously said
  "shares" while the code duplicated.

- Two rows added to the predicate table ("README.md", "_drafts") to kill the
  extension-check mutant, and the ADR's "the first version hand-listed 9 of 29"
  corrected to "the pre-glob version" (the first version listed 5).

Verified by exit code, not by grepping a line: biome 0, typecheck 0, vitest 0
(1143 passing). The orphan gate and the drop-`$` analysis both bite-proved.
